BrdData_example

NEED EXPL

Usage

BrdData_example

Format

list of "raster" with the values of a raster in a matrix; and "metadata" with the extent and resolution to build back the raster in memory.

matrix

matrix of pixel values

metadata

List of metadata: extent, resolution, and CRS.

Source

NEED EXPL